moving normalizeConfig to its own util file
