+++ paywall/src/unlock.js/startup.ts
@@ -6,32 +6,7 @@ import StartupConstants from './startupTypes'
 import { walletStatus } from '../utils/wallet'
 import { checkoutHandlerInit } from './postMessageHub'
 import { PostMessages } from '../messageTypes'
-
-/**
- * convert all of the lock addresses to lower-case so they are normalized across the app
- */
-export function normalizeConfig(unlockConfig: any) {
-  if (
-    !unlockConfig ||
-    !unlockConfig.locks ||
-    typeof unlockConfig.locks !== 'object'
-  )
-    return unlockConfig
-  const lockAddresses = Object.keys(unlockConfig.locks)
-  if (!lockAddresses.length) {
-    return unlockConfig
-  }
-  const normalizedConfig = {
-    ...unlockConfig,
-    locks: lockAddresses.reduce((allLocks, address) => {
-      return {
-        ...allLocks,
-        [address.toLowerCase()]: unlockConfig.locks[address],
-      }
-    }, {}),
-  }
-  return normalizedConfig
-}
+import { normalizeConfig } from '../utils/config'
 
 // Temporary helper to dispatch locked event when we fail early
 function dispatchEvent(detail: any, window: any) {
